docs(readme): update instructions to use terragrunt commands
instead of creating a new workspace in terraform and then running a plan against different environments. this can instead be done using `terragrunt`. this has been done to reduce repetitiveness as each environment directory needs to contain a `main.tf`, `module.tf` and `variables.tf` and terragrunt simply copies these files to the directory when required.